requirements.txt
langchain-google-genai
langchain-community
pillow
google-generativeai
tiktoken
langchain
flask
GitPython
python-dotenv
